todo-app-phase-3/frontend/backend_core/requirements.txt
fastapi
uvicorn
sqlmodel
openai
python-dotenv
asyncpg
pydantic
